modifying fetch_vcf to work for both apollo and jbrowse

+++ scripts/fetch_vcf.sh
@@ -1,27 +1,41 @@
 #!/bin/bash
+
+# either jbrowse or apollo
+CONTEXT=$1
+
+WORMVCF="VCF_WBcel235_20.vcf.gz"
+FLYVCF="VCF_GRCz11_22.vcf.gz"
+ZEBRAFISHVCF="VCF_GRCz11_22.vcf.gz" 
+MOUSEVCF="VCF_GRCm38_21.vcf.gz"
+RATVCF="VCF_Rnor60_19.vcf.gz"
+
+JBROWSE=""
+if [ "$CONTEXT" == "apollo" ]; then
+JBROWSE="jbrowse"
 rm -rf ../jbrowse/data/worm/jbrowse
 mkdir ../jbrowse/data/worm/jbrowse
-wget -P ../jbrowse/data/worm/jbrowse https://s3.amazonaws.com/agrjbrowse/VCF/VCF_WBcel235_20.vcf.gz
-wget -P ../jbrowse/data/worm/jbrowse https://s3.amazonaws.com/agrjbrowse/VCF/VCF_WBcel235_20.vcf.gz.tbi
-
-
 rm -rf ../jbrowse/data/zebrafish/jbrowse
 mkdir  ../jbrowse/data/zebrafish/jbrowse
-wget -P ../jbrowse/data/zebrafish/jbrowse https://s3.amazonaws.com/agrjbrowse/VCF/VCF_GRCz11_22.vcf.gz
-wget -P ../jbrowse/data/zebrafish/jbrowse https://s3.amazonaws.com/agrjbrowse/VCF/VCF_GRCz11_22.vcf.gz.tbi
-
 rm -rf ../jbrowse/data/fly/jbrowse
 mkdir  ../jbrowse/data/fly/jbrowse
-wget -P ../jbrowse/data/fly/jbrowse https://s3.amazonaws.com/agrjbrowse/VCF/VCF_R6_20.vcf.gz
-wget -P ../jbrowse/data/fly/jbrowse https://s3.amazonaws.com/agrjbrowse/VCF/VCF_R6_20.vcf.gz.tbi
-
 rm -rf ../jbrowse/data/MGI/jbrowse
 mkdir  ../jbrowse/data/MGI/jbrowse
-wget -P ../jbrowse/data/MGI/jbrowse https://s3.amazonaws.com/agrjbrowse/VCF/VCF_GRCm38_21.vcf.gz
-wget -P ../jbrowse/data/MGI/jbrowse https://s3.amazonaws.com/agrjbrowse/VCF/VCF_GRCm38_21.vcf.gz.tbi
-
 rm -rf ../jbrowse/data/RGD/jbrowse
 mkdir  ../jbrowse/data/RGD/jbrowse
-wget -P ../jbrowse/data/RGD/jbrowse https://s3.amazonaws.com/agrjbrowse/VCF/VCF_Rnor60_19.vcf.gz
-wget -P ../jbrowse/data/RGD/jbrowse https://s3.amazonaws.com/agrjbrowse/VCF/VCF_Rnor60_19.vcf.gz.tbi
+fi
+
+wget -P ../jbrowse/data/worm/$JBROWSE https://s3.amazonaws.com/agrjbrowse/VCF/$WORMVCF
+wget -P ../jbrowse/data/worm/$JBROWSE https://s3.amazonaws.com/agrjbrowse/VCF/$WORMVCF.tbi
+
+wget -P ../jbrowse/data/zebrafish/$JBROWSE https://s3.amazonaws.com/agrjbrowse/VCF/$ZEBRAFISHVCF
+wget -P ../jbrowse/data/zebrafish/$JBROWSE https://s3.amazonaws.com/agrjbrowse/VCF/$ZEBRAFISHVCF.tbi
+
+wget -P ../jbrowse/data/fly/$JBROWSE https://s3.amazonaws.com/agrjbrowse/VCF/$FLYVCF
+wget -P ../jbrowse/data/fly/$JBROWSE https://s3.amazonaws.com/agrjbrowse/VCF/$FLYVCF.tbi
+
+wget -P ../jbrowse/data/MGI/$JBROWSE https://s3.amazonaws.com/agrjbrowse/VCF/$MOUSEVCF
+wget -P ../jbrowse/data/MGI/$JBROWSE https://s3.amazonaws.com/agrjbrowse/VCF/$MOUSEVCF.tbi
+
+wget -P ../jbrowse/data/RGD/$JBROWSE https://s3.amazonaws.com/agrjbrowse/VCF/$RATVCF
+wget -P ../jbrowse/data/RGD/$JBROWSE https://s3.amazonaws.com/agrjbrowse/VCF/$MOUSEVCF.tbi
